+++ include/gostone/keymgmt/keymgmt_params.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <openssl/ec.h>
+
+EC_GROUP* GsGetEcGroup( const OSSL_PARAM* param );
