+++ provider/encode/encode_to_text.c
@@ -7,7 +7,7 @@
 #include <openssl/err.h>
 
 #include <gostone/provider_bio.h>
-#include <gostone/keymgmt/asymm_key.h>
+#include <gostone/keymgmt/keymgmt_akey.h>
 #include <gostone/encode/encode_impl.h>
 #include <gostone/encode/encode_common.h>
